Add readRemoteFile.js
