+++ .gitignore
@@ -7,7 +7,6 @@ node_modules
 build
 dist
 
-
 # misc
 .DS_Store
 .env
