+++ src/index.js
@@ -3,6 +3,7 @@ import PapaParse from 'papaparse'
 export { readString } from './readString'
 export { default as CSVReader } from './CSVReader'
 export { jsonToCSV } from './jsonToCSV'
+export { readRemoteFile } from './readRemoteFile'
 
 export const BAD_DELIMITERS = PapaParse.BAD_DELIMITERS
 export const RECORD_SEP = PapaParse.RECORD_SEP
