+++ src/readRemoteFile.js
@@ -1,0 +1,5 @@
+import PapaParse from 'papaparse'
+
+export function readRemoteFile (url, options = {}) { 
+  PapaParse.parse(url, Object.assign({}, {download: true}, options));
+}
